Read and write ability; ignore course-supplied material.
Add the ability to read the contents of an input file, and the ability
to write content to another file of the same base name as the input
file.
Calling "hack_assembler abc.asm" will now create a file abc.hack with the
same content as abc.asm

Also update .gitignore to ignore content supplied by the Coursera course
as it is not mine to share/upload/etc.