Write preprocessor; Add skeleton functions; Cleanup comments
Wrote the line preprocessor, which cuts out all comments and whitespace
in order to transform the content into a form the parser understands.
The preprocessor returns either None or Some(String) where the string is
something the parser can successfully break down. Lines that are just
comments as well as whitespace (empty lines) become None, anything else
becomes Some(String). The entire program assumes the assembly input is
correct and so no validation is performed.

Add skeleton functions with todo!() or unimplimented!() that are yet to
be written so the program can run as-is.

Cleaned up some comments that were no longer needed or relevant.