Write translator and symbol handling
Write the translator, and give the program the ability to handle
non-numeric A commands and labels. The assembler works in two phases,
the first of which has two passes. In the first phase, we populate the
symbol table. The first pass inserts only the line number of any loop
declarations, while the second assigns an address for any variables in
the program. The second phase translates each line into the binary, line
by line.